feat: Fill out the load_synthetic_dataset function
This function assumes that the dataset is loaded into a particular
directory in the resources file. It randomly loads images from the
dataset, and converts them using import_tiff, and then returns the
images and labels as (training_images, training_labels), (test_images,
test_labels). Some additional tests have been added, to check the basic
functionality. One test assumes that three particular images will be
selected (as it sets a seed value), which may cause issues on other
clients. We can deal with this issue if it araises.