feat: Add Simple CNN implementation
This is a function, if messy implementation of a cell counter. Running
it directly with `python3 cell_counter/cnn_cellcounter.py`, after about
a minute, the model is evaluated to have a mean squared error of about
169 on the test data. This implies that it is on average off by about 13
cells. Some tests have been implemented, but it's rather difficult to
think of good tests for the neural network that don't just examine the
tensorflow library.