feat: Add functions for filtering imported images
Four function have been added. The get function queries the specified
path for the images and their metadata, the save function saves the
dataframe as a csv to the specified location, the load function loads
the dataframe, and the final function uses a given dataframe to import
the actual images and labels.